pyproject.toml
[project]
name = "ccta-net"
version = "0.1.0"
description = "Cardiac Computed Tomography Angiography Plane Prediction and Comprehensive LV Segmentation"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "dash[compress]>=2.18.1",
    "kaleido==0.2.0",
    "mkdocs-material>=9.5.44",
    "mkdocs>=1.6.1",
    "mkdocstrings[python]>=0.27.0",
    "monai-weekly>=1.4.dev2437",
    "nibabel>=5.2.1",
    "numpy>=1.26.4",
    "pandas>=2.2.3",
    "plotly>=5.24.1",
    "pydantic>=2.9.2",
    "roma>=1.5.0",
    "scikit-image>=0.24.0",
    "tomlkit>=0.13.2",
    "torch>=2.4.1",
    "tqdm>=4.66.5",
]

[tool.uv.sources]
